+++ Source/Internal/HAConnectionImpl.swift
@@ -59,6 +59,7 @@ internal class HAConnectionImpl: HAConnection {
     let responseController: HAResponseController
     let reconnectManager: HAReconnectManager
     var connectAutomatically: Bool
+    var retrySubscriptionTokens: [HACancellable] = []
     private(set) lazy var caches: HACachesContainer = .init(connection: self)
 
     init(
@@ -81,6 +82,19 @@ internal class HAConnectionImpl: HAConnection {
         reconnectManager.delegate = self
     }
 
+    private func setupResubscribeEvents() {
+        guard retrySubscriptionTokens.isEmpty else { return }
+
+        retrySubscriptionTokens = requestController.retrySubscriptionsEvents.map { [requestController] event in
+            commonSubscribe(
+                to: .events(event),
+                allowConnecting: false,
+                initiated: nil,
+                handler: { _, _ in requestController.retrySubscriptions() }
+            )
+        }
+    }
+
     // MARK: - Connection Handling
 
     public func connect() {
@@ -101,6 +115,8 @@ internal class HAConnectionImpl: HAConnection {
             disconnect(permanently: false, error: ConnectError.noConnectionInfo)
             return
         }
+
+        setupResubscribeEvents()
 
         let connection: WebSocket = {
             guard let existing = self.connection else {
@@ -188,12 +204,17 @@ internal class HAConnectionImpl: HAConnection {
 
     private func commonSubscribe(
         to request: HARequest,
+        allowConnecting: Bool = true,
         initiated: SubscriptionInitiatedHandler?,
         handler: @escaping SubscriptionHandler
     ) -> HACancellable {
         let sub = HARequestInvocationSubscription(request: request, initiated: initiated, handler: handler)
         requestController.add(sub)
-        defer { connectAutomaticallyIfNeeded() }
+        defer {
+            if allowConnecting {
+                connectAutomaticallyIfNeeded()
+            }
+        }
         return HACancellableImpl { [requestController] in
             requestController.cancel(sub)
         }
@@ -201,21 +222,27 @@ internal class HAConnectionImpl: HAConnection {
 
     private func commonSubscribe<T>(
         to request: HATypedSubscription<T>,
+        allowConnecting: Bool = true,
         initiated: SubscriptionInitiatedHandler?,
         handler: @escaping (HACancellable, T) -> Void
     ) -> HACancellable {
-        commonSubscribe(to: request.request, initiated: initiated, handler: { [workQueue, callbackQueue] token, data in
-            workQueue.async {
-                do {
-                    let value = try T(data: data)
-                    callbackQueue.async {
-                        handler(token, value)
+        commonSubscribe(
+            to: request.request,
+            allowConnecting: allowConnecting,
+            initiated: initiated,
+            handler: { [workQueue, callbackQueue] token, data in
+                workQueue.async {
+                    do {
+                        let value = try T(data: data)
+                        callbackQueue.async {
+                            handler(token, value)
+                        }
+                    } catch {
+                        HAGlobal.log(.info, "couldn't parse data \(error)")
                     }
-                } catch {
-                    HAGlobal.log(.info, "couldn't parse data \(error)")
                 }
             }
-        })
+        )
     }
 
     @discardableResult